requirements.txt
AppOpener
comtypes
customtkinter
edge-tts
google-generativeai
groq
mediapipe
numpy
opencv-python
Pillow
psutil
pyautogui
pycaw
pygame
pygetwindow
pyperclip
pyttsx3
pywhatkit
requests
speedtest-cli
SpeechRecognition
python-dotenv
ytmusicapi
pyaudio
pytest